# Optimize requeue during flow control conditions