@leftwo 2023-09-26
I wonder if it's time (unrelated to this PR specifically) to increase the default count for crutest perf.  Things are so much
faster now and having 3 seconds of IO might not be long enough.